cleanup incrementing [ci skip]

+++ src/permutations.h
@@ -5,16 +5,16 @@
 typedef struct __permutations permutations;
 typedef struct __permutations *permutations_t;
 
-permutations_t perm_init(unsigned int input_len, const char* alphabet);
+permutations_t perm_init(size_t input_len, const char* alphabet);
 
 /*  Free all the memory */
 void perm_free(permutations_t p);
 
 void perm_inc(permutations_t p);
-void perm_add(permutations_t p, unsigned int num);
+void perm_add(permutations_t p, size_t num);
 
 void perm_dec(permutations_t p);
-void perm_sub(permutations_t p, unsigned int num);
+void perm_sub(permutations_t p, size_t num);
 
 
 const unsigned short* perm_current_permutation(permutations_t p);
